+++ fidget-shapes/src/lib.rs
@@ -31,8 +31,8 @@ pub struct Circle {
     #[facet(default = Vec2::new(0.0, 0.0))]
     pub center: Vec2,
     /// Circle radius
-    #[facet(default = 1.0)]
-    pub radius: f64,
+    #[facet(default = 1.0f32)]
+    pub radius: f32,
 }
 
 impl From<Circle> for Tree {
@@ -71,8 +71,8 @@ pub struct Sphere {
     #[facet(default = Vec3::new(0.0, 0.0, 0.0))]
     pub center: Vec3,
     /// Sphere radius
-    #[facet(default = 1.0)]
-    pub radius: f64,
+    #[facet(default = 1.0f32)]
+    pub radius: f32,
 }
 
 impl From<Sphere> for Tree {
@@ -121,7 +121,7 @@ impl From<Union> for Tree {
     fn from(v: Union) -> Self {
         if v.input.is_empty() {
             // XXX should this be an error instead?
-            Tree::constant(f64::INFINITY)
+            Tree::constant(f32::INFINITY)
         } else {
             fn recurse(s: &[Tree]) -> Tree {
                 match s.len() {
@@ -146,7 +146,7 @@ pub struct Blend {
     /// Second shape input
     pub b: Tree,
     /// Blending radius
-    pub radius: f64,
+    pub radius: f32,
 }
 
 impl From<Blend> for Tree {
@@ -174,7 +174,7 @@ impl From<Intersection> for Tree {
     fn from(v: Intersection) -> Self {
         if v.input.is_empty() {
             // XXX should this be an error instead?
-            Tree::constant(-f64::INFINITY)
+            Tree::constant(-f32::INFINITY)
         } else {
             fn recurse(s: &[Tree]) -> Tree {
                 match s.len() {
@@ -231,7 +231,7 @@ pub struct Move {
 impl From<Move> for Tree {
     fn from(v: Move) -> Self {
         v.shape.remap_affine(nalgebra::convert(
-            nalgebra::Translation3::<f64>::new(
+            nalgebra::Translation3::<f32>::new(
                 -v.offset.x,
                 -v.offset.y,
                 -v.offset.z,
@@ -253,7 +253,7 @@ pub struct Scale {
 impl From<Scale> for Tree {
     fn from(v: Scale) -> Self {
         v.shape
-            .remap_affine(nalgebra::convert(nalgebra::Scale3::<f64>::new(
+            .remap_affine(nalgebra::convert(nalgebra::Scale3::<f32>::new(
                 1.0 / v.scale.x,
                 1.0 / v.scale.y,
                 1.0 / v.scale.z,
@@ -267,15 +267,15 @@ pub struct ScaleUniform {
     /// Shape to scale
     pub shape: Tree,
     /// Scale to apply
-    #[facet(default = 1.0)]
-    pub scale: f64,
+    #[facet(default = 1.0f32)]
+    pub scale: f32,
 }
 
 impl From<ScaleUniform> for Tree {
     fn from(v: ScaleUniform) -> Self {
         let s = 1.0 / v.scale;
         v.shape
-            .remap_affine(nalgebra::convert(nalgebra::Scale3::<f64>::new(
+            .remap_affine(nalgebra::convert(nalgebra::Scale3::<f32>::new(
                 s, s, s,
             )))
     }
@@ -317,8 +317,8 @@ pub struct ReflectX {
     pub shape: Tree,
 
     /// X offset
-    #[facet(default = 0.0)]
-    pub offset: f64,
+    #[facet(default = 0.0f32)]
+    pub offset: f32,
 }
 
 impl From<ReflectX> for Tree {
@@ -341,8 +341,8 @@ pub struct ReflectXY {
     pub shape: Tree,
 
     /// Plane about which to reflect the shape
-    #[facet(default = 0.0)]
-    pub offset: f64,
+    #[facet(default = 0.0f32)]
+    pub offset: f32,
 }
 
 impl From<ReflectXY> for Tree {
@@ -365,8 +365,8 @@ pub struct ReflectY {
     pub shape: Tree,
 
     /// Y offset
-    #[facet(default = 0.0)]
-    pub offset: f64,
+    #[facet(default = 0.0f32)]
+    pub offset: f32,
 }
 
 impl From<ReflectY> for Tree {
@@ -389,8 +389,8 @@ pub struct ReflectZ {
     pub shape: Tree,
 
     /// Z offset
-    #[facet(default = 0.0)]
-    pub offset: f64,
+    #[facet(default = 0.0f32)]
+    pub offset: f32,
 }
 
 impl From<ReflectZ> for Tree {
@@ -417,8 +417,8 @@ pub struct Rotate {
     pub axis: Axis,
 
     /// Angle to rotate (in degrees)
-    #[facet(default = 0.0)]
-    pub angle: f64,
+    #[facet(default = 0.0f32)]
+    pub angle: f32,
 
     /// Center of rotation
     #[facet(default = Vec3::new(0.0, 0.0, 0.0))]
@@ -434,7 +434,7 @@ impl From<Rotate> for Tree {
         let d = -v.angle.to_radians();
         let axis = v.axis.vec();
         let shape = shape.remap_affine(nalgebra::convert(
-            nalgebra::Rotation3::<f64>::new(nalgebra::Vector3::from(d * *axis)),
+            nalgebra::Rotation3::<f32>::new(nalgebra::Vector3::from(d * *axis)),
         ));
         Move {
             shape,
@@ -451,8 +451,8 @@ pub struct RotateX {
     pub shape: Tree,
 
     /// Angle to rotate (in degrees)
-    #[facet(default = 0.0)]
-    pub angle: f64,
+    #[facet(default = 0.0f32)]
+    pub angle: f32,
 
     /// Center of rotation
     #[facet(default = Vec3::new(0.0, 0.0, 0.0))]
@@ -478,8 +478,8 @@ pub struct RotateY {
     pub shape: Tree,
 
     /// Angle to rotate (in degrees)
-    #[facet(default = 0.0)]
-    pub angle: f64,
+    #[facet(default = 0.0f32)]
+    pub angle: f32,
 
     /// Center of rotation
     #[facet(default = Vec3::new(0.0, 0.0, 0.0))]
@@ -505,8 +505,8 @@ pub struct RotateZ {
     pub shape: Tree,
 
     /// Angle to rotate (in degrees)
-    #[facet(default = 0.0)]
-    pub angle: f64,
+    #[facet(default = 0.0f32)]
+    pub angle: f32,
 
     /// Center of rotation
     #[facet(default = Vec3::new(0.0, 0.0, 0.0))]
@@ -533,8 +533,8 @@ pub struct RevolveY {
     /// Shape to revolve
     pub shape: Tree,
     /// X offset about which to revolve
-    #[facet(default = 0.0)]
-    pub offset: f64,
+    #[facet(default = 0.0f32)]
+    pub offset: f32,
 }
 
 impl From<RevolveY> for Tree {
@@ -557,11 +557,11 @@ pub struct ExtrudeZ {
     /// Shape to extrude
     pub shape: Tree,
     /// Lower bounds of the extrusion
-    #[facet(default = 0.0)]
-    pub lower: f64,
+    #[facet(default = 0.0f32)]
+    pub lower: f32,
     /// Upper bounds of the extrusion
-    #[facet(default = 1.0)]
-    pub upper: f64,
+    #[facet(default = 1.0f32)]
+    pub upper: f32,
 }
 
 impl From<ExtrudeZ> for Tree {
@@ -580,11 +580,11 @@ pub struct LoftZ {
     /// Upper shape
     pub b: Tree,
     /// Lower bounds of the loft
-    #[facet(default = 0.0)]
-    pub lower: f64,
+    #[facet(default = 0.0f32)]
+    pub lower: f32,
     /// Upper bounds of the loft
-    #[facet(default = 1.0)]
-    pub upper: f64,
+    #[facet(default = 1.0f32)]
+    pub upper: f32,
 }
 
 impl From<LoftZ> for Tree {
@@ -607,11 +607,11 @@ pub struct RepeatX {
     /// Shape to repeat
     pub shape: Tree,
     /// Radius of the region to repeat
-    #[facet(default = 1.0)]
-    pub radius: f64,
+    #[facet(default = 1.0f32)]
+    pub radius: f32,
     /// X position about which to repeat
-    #[facet(default = 0.0)]
-    pub offset: f64,
+    #[facet(default = 0.0f32)]
+    pub offset: f32,
 }
 
 impl From<RepeatX> for Tree {
@@ -805,7 +805,7 @@ mod test {
         #[derive(Clone, facet::Facet)]
         struct BadShape {
             #[facet(default)]
-            center: f64,
+            center: f32,
         }
         impl From<BadShape> for Tree {
             fn from(_: BadShape) -> Tree {
@@ -850,7 +850,7 @@ mod test {
 
         // Check the three repetitions closest to the center
         for i in 0..=1000 {
-            let x = (i as f64 / 1000.0) - (1000.0 - i as f64) / 1000.0;
+            let x = (i as f32 / 1000.0) - (1000.0 - i as f32) / 1000.0;
             let vc = ctx.eval_xyz(c, x, 0.0, 0.0).unwrap();
             let err = (vc - ctx.eval_xyz(r, x, 0.0, 0.0).unwrap()).abs();
             assert!(err < 1e-6, "bad err {err} at {x}");
@@ -876,7 +876,7 @@ mod test {
 
         // Check the three repetitions closest to the center
         for i in 0..=1000 {
-            let x = 0.75 * ((i as f64 / 1000.0) - (1000.0 - i as f64) / 1000.0)
+            let x = 0.75 * ((i as f32 / 1000.0) - (1000.0 - i as f32) / 1000.0)
                 + 1.0;
             let vc = ctx.eval_xyz(c, x, 0.0, 0.0).unwrap();
             let err = (vc - ctx.eval_xyz(r, x, 0.0, 0.0).unwrap()).abs();
